Move transformOrderItems to helpers folder
AB#7884
