+++ app/pages/preview/helpers/transformOrderItems.js
